fix(cx): return a consistent string type
This has a downside that if you call cx with values that are all falsy, an empty class will be written, but upside that you can rely on string manipulation later if needed.

This changes the snapshot of highlight, but not the real life usage, as that's called with a string that definitely exists.

This isn't a breaking change in my opinion, as an empty className and no className act identical, and this solution is much simpler than #20 or changing all code related to classNames in InstantSearch to have every key optional (even if they actually aren't going to be undefined)

For reference, the `classnames` package is also typed to always return a string.